requirements.txt
docx==0.2.4
fitz==0.0.1.dev2
openai==0.27.8
Pillow==10.3.0
tqdm==4.65.0
